+++ Huppy/Huppy.Kernel/Abstraction/IRepository.cs
